structure of inventory pyfilesystem
